requirements.txt
discord~=2.3.2
python-dotenv~=1.0.1
pillow~=10.3.0
feedparser~=6.0.11